// src/models/Onboarding.ts
export class SignupDetails { 
    brand_id:string; 
    domain_id:string; 
    category_id: string;    
    category: string;
    brand_name: string;
    about: string;
    country: string;
    email_id: string;
    phone_number: number;
    country_code: string;
    street: string;
    city: string;
    state: string;
    postal_code: string;
    authorised_user_name: string;
    authorised_user_designation: string;
    authorised_user_email_id: string;
    registered_business_name: string;
    registered_type: string;
    kyc_doc_type: string;
    kyc_doc_number: string;
    kyc_doc_url: string;
    brand_image_url: string;
    authorised_user_password: string;
  }



  